create “INSTALLING” section in README for the most simple instructions
Currently the most simple `npm install sqlite3` is buried in the most deep bottom of the “BUILDING” section. And the instructions for building the `master` branch immediately precede it.

This commit creates a separate “INSTALLING” section (placed before “BUILDING”) containing only the most simple instructions.

A new method of building the `master` branch is introduced. The user's system won't even need `git` for that because a tarball is downloaded by npm from GitHub.

A message about the necessity of meeting node-gyp's requirements is given.